fix: recipientId must be Long

+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/data/database/model/MessageDataDbModel.kt
@@ -13,7 +13,7 @@ data class MessageDataDbModel(
     val streamId: Long,
     val senderId: Long,
     val content: String,
-    val recipientId: Int,
+    val recipientId: Long,
     val timestamp: Long,
     val subject: String,
     val isMeMessage: Boolean,
